Move voxel cone trace stage to its own file
