+++ src/cone_tracing/mod.rs
@@ -1,0 +1,2 @@
+mod voxel_cone_trace;
+pub use voxel_cone_trace::voxel_cone_trace;
